Postgres: Moving to harness object
